Added the source code for Blog post 1
Blog post 1 - Capturing performance counters to optimize SQL Server performance on AWS EC2 instances is the first in a series of Power BI for the SQL Server DBA on how to visualize SQL Server metrics to make things that matter pop. In this script, I show which performance counters I've found useful for modifying EC2 EBS storage to optimize for cost and/or performance for SQL Server workloads.